+++ PSPOS.ApiService/Repositories/OrderRepository.cs
@@ -130,38 +130,6 @@ public class OrderRepository : IOrderRepository
         return await _context.Transactions.FindAsync(transactionId);
     }
 
-    public async Task<Giftcard?> GetGiftCardByCode(string giftcardCode)
-    {
-        if (string.IsNullOrWhiteSpace(giftcardCode))
-            throw new ArgumentException("Gift card code cannot be null or empty.", nameof(giftcardCode));
-
-        // Query to find the gift card object with the matching code
-        var giftCard = await _context.GiftCards
-            .FirstOrDefaultAsync(gc => gc.Code == giftcardCode);
-
-        return giftCard; // Returns null if no matching gift card is found
-    }
-
-    public async Task UpdateGiftCardAmountAsync(Giftcard giftcard)
-    {
-        if (giftcard == null)
-            throw new ArgumentNullException(nameof(giftcard), "Gift card object cannot be null.");
-
-        // Find the existing gift card in the database by its ID or Code
-        var existingGiftCard = await _context.GiftCards
-            .FirstOrDefaultAsync(gc => gc.Code == giftcard.Code);
-
-        if (existingGiftCard == null)
-            throw new InvalidOperationException("Gift card not found.");
-
-        // Update the gift card's amount
-        existingGiftCard.Amount = giftcard.Amount;
-
-        // Save the changes to the database
-        _context.GiftCards.Update(existingGiftCard);
-        await _context.SaveChangesAsync();
-    }
-
     public async Task<IEnumerable<Payment>> GetAllPaymentsOfTransactionAsync(Guid transactionId)
     {
         return await _context.Payments.Where(p => p.TransactionId == transactionId).ToArrayAsync();
@@ -270,7 +238,16 @@ public class OrderRepository : IOrderRepository
         {
             foreach (var productOrServiceId in group.productOrServiceIds ?? Array.Empty<Guid>())
             {
-                // Add both ProductId and ServiceId mappings
+                if (!productGroupDict.ContainsKey(productOrServiceId))
+                {
+                    productGroupDict[productOrServiceId] = group.Id;
+                }
+            }
+        }
+        foreach (var group in serviceGroups)
+        { 
+            foreach (var productOrServiceId in group.productOrServiceIds ?? Array.Empty<Guid>())
+            {
                 if (!productGroupDict.ContainsKey(productOrServiceId))
                 {
                     productGroupDict[productOrServiceId] = group.Id;
